+++ src/tests/basic.test.ts
@@ -21,10 +21,10 @@ test("basic: happy path with no payload", async () => {
 
   const mockReq = new MockReq()
     .setMethod("POST")
-    .setUrl("/api/procedures")
+    .setUrl("/api/procedures/helloWorld")
     .setHeader("content-type", "application/json")
     .setHeader("host", "localhost")
-    .setBody(JSON.stringify({ procedure: "helloWorld", payload: null }));
+    .setBody("null");
 
   const mockRes = new MockRes();